change motionDetector from async to time sleep
as usual it's my own folly that foils my plans.
asyncio.sleep() returns immideately, and allows the next thing to run.

What I want to do in the script is not to run many things at once,
I want to run a few things in series, with lots of downtime in between.